refactor wepo
